Added event listener to add btn
