+++ index.html
@@ -33,7 +33,7 @@
         <form action="">
             <input type="text" id="title" placeholder="title">
             <input type="text" id="author" placeholder="author"> 
-            <div id="formbut"><button id="addbtn" onclick="a.addBooks()"> Add</button></div>
+            <div id="formbut"><button id="addbtn"> Add</button></div>
         </form>
        </div>  
     </section>
@@ -59,6 +59,6 @@
     <p>Copyright</p>
  </footer>
     
-    <script src="./index.js"></script>
+    <script type="module" src="./index.js"></script>
 </body>
 </html>
