--- modules/booksArr.js
@@ -1,14 +1,0 @@
-const booksArr = [
-  {
-    Id: 1,
-    title: 'Lorem ipsum',
-    author: 'Testeroo Testyy',
-  },
-  {
-    Id: 2,
-    title: 'Second Books',
-    author: 'Testeroo Testyy',
-  },
-];
-
-export default { booksArr };
